# Conflicts:
#	pmm-app/src/pmm-dbaas/components/AddClusterButton/AddClusterButton.tsx
#	pmm-app/src/pmm-dbaas/components/DBCluster/AddDBClusterModal/AddDBClusterModal.tsx
#	pmm-app/src/pmm-dbaas/components/DBCluster/AddDBClusterModal/DBClusterAdvancedOptions/DBClusterAdvancedOptions.tsx
#	pmm-app/src/pmm-dbaas/components/DBCluster/DBCluster.test.tsx
#	pmm-app/src/pmm-dbaas/components/DBCluster/DBClusterActions/DBClusterActions.test.tsx
#	pmm-app/src/pmm-dbaas/components/DBCluster/EditDBClusterModal/DBClusterAdvancedOptions/DBClusterAdvancedOptions.tsx
#	pmm-app/src/pmm-dbaas/components/DBCluster/__mocks__/dbClustersStubs.ts
#	pmm-app/src/pmm-dbaas/components/Kubernetes/KubernetesInventory.tsx
#	pmm-app/tests/pages/homePage.js

+++ pmm-app/src/pmm-dbaas/components/Kubernetes/KubernetesInventory.tsx
@@ -5,10 +5,13 @@ import { Table } from 'shared/components/Elements/Table/Table';
 import { Messages } from 'pmm-dbaas/DBaaS.messages';
 import { Form, FormRenderProps } from 'react-final-form';
 import { Modal } from 'shared/components/Elements/Modal/Modal';
+import { CheckboxField, FormElement } from 'shared/components/Form';
+import { Databases } from 'shared/core';
 import { getStyles } from './Kubernetes.styles';
-import { Kubernetes, NewKubernetesCluster, KubernetesProps } from './Kubernetes.types';
+import { NewKubernetesCluster, KubernetesProps } from './Kubernetes.types';
 import { AddClusterButton } from '../AddClusterButton/AddClusterButton';
-import { CheckboxField, FormElement } from '../../../shared/components/Form';
+import { OperatorStatusItem } from './OperatorStatusItem/OperatorStatusItem';
+import { KubernetesClusterStatus } from './KubernetesClusterStatus/KubernetesClusterStatus';
 
 export const KubernetesInventory: FC<KubernetesProps> = ({
   kubernetes,
@@ -17,7 +20,9 @@ export const KubernetesInventory: FC<KubernetesProps> = ({
   loading,
 }) => {
   const styles = useStyles(getStyles);
-  const [kubernetesToDelete, setKubernetesToDelete] = useState<Kubernetes>({ kubernetesClusterName: '' });
+  const [kubernetesToDelete, setKubernetesToDelete] = useState<any>({
+    kubernetesClusterName: '',
+  });
   const [deleteModalVisible, setDeleteModalVisible] = useState(false);
   const [addModalVisible, setAddModalVisible] = useState(false);
   const { required } = validators;
@@ -25,6 +30,19 @@ export const KubernetesInventory: FC<KubernetesProps> = ({
     {
       Header: Messages.kubernetes.table.nameColumn,
       accessor: 'kubernetesClusterName',
+    },
+    {
+      Header: Messages.kubernetes.table.clusterStatusColumn,
+      accessor: (element) => <KubernetesClusterStatus status={element.status} />,
+    },
+    {
+      Header: Messages.kubernetes.table.operatorsStatusColumn,
+      accessor: (element) => (
+        <div>
+          <OperatorStatusItem databaseType={Databases.mysql} status={element.operators.xtradb.status} />
+          <OperatorStatusItem databaseType={Databases.mongodb} status={element.operators.psmdb.status} />
+        </div>
+      ),
     },
     {
       Header: Messages.kubernetes.table.actionsColumn,
@@ -135,8 +153,10 @@ export const KubernetesInventory: FC<KubernetesProps> = ({
                     variant="destructive"
                     size="md"
                     onClick={() => {
-                      deleteKubernetes(kubernetesToDelete, Boolean(form.getState().values.force));
-                      setDeleteModalVisible(false);
+                      if (kubernetesToDelete) {
+                        deleteKubernetes(kubernetesToDelete, Boolean(form.getState().values.force));
+                        setDeleteModalVisible(false);
+                      }
                     }}
                     data-qa="delete-kubernetes-button"
                   >
